+++ adcio_placement/src/main/java/ai/corca/adcio_placement/mapper/BannerMapper.kt
@@ -1,12 +1,12 @@
 package ai.corca.adcio_placement.mapper
 
-import ai.corca.adcio_placement.model.Banner
-import ai.corca.adcio_placement.network.data.BannerData
+import ai.corca.adcio_placement.model.banner.Banner
+import ai.corca.adcio_placement.network.data.response.banner.BannerData
 
 internal fun BannerData.toBanner(): Banner = Banner(
     id = this.id,
     clientId = this.clientId,
-    imageUrl = this.imageUrl,
+    creative = this.creative,
     name = this.name,
     url = this.url,
     categoryId = this.categoryId,
@@ -21,7 +21,7 @@ internal fun BannerData.toBanner(): Banner = Banner(
 internal fun Banner.toBannerData(): BannerData = BannerData(
     id = this.id,
     clientId = this.clientId,
-    imageUrl = this.imageUrl,
+    creative = this.creative,
     name = this.name,
     url = this.url,
     categoryId = this.categoryId,