+++ adcio_placement/src/main/java/ai/corca/adcio_placement/model/AdcioSuggestion.kt
@@ -1,7 +1,0 @@
-package ai.corca.adcio_placement.model
-
-data class AdcioSuggestion(
-    val banner: Banner?,
-    val logOptions: LogOptions,
-    val product: Product,
-)