+++ adcio_placement/src/main/java/ai/corca/adcio_placement/model/banner/AdcioSuggestionBanner.kt
@@ -1,0 +1,6 @@
+package ai.corca.adcio_placement.model.banner
+
+data class AdcioSuggestionBanner(
+    val banner: Banner?,
+    val logOptions: BannerLogOptions,
+)