+++ adcio_placement/src/main/java/ai/corca/adcio_placement/model/product/AdcioSuggestionProduct.kt
@@ -1,0 +1,6 @@
+package ai.corca.adcio_placement.model.product
+
+data class AdcioSuggestionProduct(
+    val logOptions: ProductLogOptions,
+    val product: Product,
+)