+++ adcio_placement/src/main/java/ai/corca/adcio_placement/network/PlacementUrl.kt
@@ -5,6 +5,9 @@ internal object PlacementUrl {
     var baseUrl: String = "https://api.adcio.ai/"
 
     object EndPoint {
-        const val suggestions = "suggestions"
+        const val recProducts = "recommendations/products"
+        const val adBanners = "advertisements/banners"
+        const val adProducts = "advertisements/products"
+        const val recBanners = "recommendations/banners"
     }
 }