+++ adcio_placement/src/main/java/ai/corca/adcio_placement/network/data/ErrorResponse.kt
@@ -1,8 +1,0 @@
-package ai.corca.adcio_placement.network.data
-
-import com.google.gson.annotations.SerializedName
-
-data class ErrorResponse(
-    @SerializedName("statusCode") val statusCode: Int,
-    @SerializedName("message") val message: Any,
-)