+++ adcio_placement/src/main/java/ai/corca/adcio_placement/network/data/LogOptionsData.kt
@@ -1,8 +1,0 @@
-package ai.corca.adcio_placement.network.data
-
-import com.google.gson.annotations.SerializedName
-
-data class LogOptionsData(
-    @SerializedName("requestId") val requestId: String,
-    @SerializedName("adsetId") val adsetId: String,
-)