+++ adcio_placement/src/main/java/ai/corca/adcio_placement/network/data/response/banner/AdcioSuggestionBannerData.kt
@@ -1,0 +1,8 @@
+package ai.corca.adcio_placement.network.data.response.banner
+
+import com.google.gson.annotations.SerializedName
+
+data class AdcioSuggestionBannerData(
+    @SerializedName("banner") val bannerData: BannerData?,
+    @SerializedName("logOptions") val logOptionsData: BannerLogOptionsData,
+)