+++ app/src/main/java/ai/corca/adcio_android_plugins/MainActivity.kt
@@ -1,10 +1,8 @@
 package ai.corca.adcio_android_plugins
 
 import ai.corca.adcio_android_plugins.agent.AgentActivity
-import ai.corca.adcio_android_plugins.analytics.AnalyticsActivity
-import ai.corca.adcio_android_plugins.core.CoreActivity
 import ai.corca.adcio_android_plugins.databinding.ActivityMainBinding
-import ai.corca.adcio_android_plugins.placement.PlacementActivity
+import ai.corca.adcio_android_plugins.suggestion.PlacementActivity
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -22,18 +20,8 @@ class MainActivity : AppCompatActivity() {
             startActivity(intent)
         }
 
-        binding.btnAnalytics.setOnClickListener {
-            val intent = Intent(this, AnalyticsActivity::class.java)
-            startActivity(intent)
-        }
-
         binding.btnPlacement.setOnClickListener {
             val intent = Intent(this, PlacementActivity::class.java)
-            startActivity(intent)
-        }
-
-        binding.btnCore.setOnClickListener {
-            val intent = Intent(this, CoreActivity::class.java)
             startActivity(intent)
         }
     }