+++ adcio_analytics/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.2
+* Feature: Change Analytics Request Scheme
+
+## 1.1.1
+* Feature: Update Api Endpoint 
+
+## 1.1.0
+* Build: Downgrade android minSDK Version
+
 ## 1.0.0
 * Update: Removed dependency on adcio core module, change ID to be managed by plugin itself