+++ adcio_placement/CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.1.1
+* Feature: Update Api Endpoint
+
+## 1.1.0
+* Build: Downgrade android minSDK Version
+
 ## 1.0.0
 * Update: Removed dependency on adcio core module, change ID to be managed by plugin itself
-* 
+
 ## 0.2.0
 * Feature: update suggest API call
 * Feature: add Error case for suggest API call